Reorganises packages again
