+++ data/duke.txt
@@ -1,0 +1,2 @@
+D | 0 | null | Sunday
+T | 0 | null
