+++ src/main/java/duke/exceptions/EmptyDescriptionException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class EmptyDescriptionException extends Exception {
     public EmptyDescriptionException(String message) {
