+++ src/main/java/duke/tasks/Task.java
@@ -1,4 +1,4 @@
-package task;
+package duke.tasks;
 
 public abstract class Task {
     protected String description;
